requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.5.0
opencv-python>=4.8.0
pyyaml>=6.0

# Model dependencies
transformers>=4.30.0
timm>=0.9.0
segment-anything>=1.0
onnxruntime>=1.15.0

# Utils
scipy>=1.10.0
matplotlib>=3.7.0
tqdm>=4.65.0
einops>=0.6.0

# Optional for training
# tensorboard>=2.12.0
# wandb>=0.14.0

# For dataset
# albumentations>=1.3.0
